feat: add Index::canonicalize method, break up test modules
Adding an Index::canonicalize method which makes ordering of files deterministic in an Index.

chore: split tests into separate module files as lib.rs was getting too long
